Set up basic command framework
